Add support for the trivia buzzer mode  (#11)
* Started adding game staate definiteions for each mode

* Added buzzer display mode constants

* Added buzzer device states

* Added documentation for diaply structs

* Added Buzzer Result data struct

* Added new virtual display functions for buzzer mode

* Added brightness buffer gradient funciton

* Removed unused function

* Updated overlay buffer to take an initial offset

* Added new buzzer mode display functions

* Added new pio environment to test sim display

* New sim display main file

* Cleanedup and added new tests

* fixed build issues

* Updated how data for specific device states is determined

* Added turn start display mode

+++ lib/InputMonitor/input_monitor.cpp
@@ -49,7 +49,7 @@ ButtonInputType ButtonInputMonitor::getAction()
     last_press_start = currentTime;
     currentPresses += 1;
 
-    logger.info("Started press ", (int)currentPresses);
+    logger.info("Started press ", currentPresses);
   }
   else if (isPressed && !value)
   {
